Move inference types to their own files
