+++ src/Elastic.Clients.Elasticsearch/Common/Infer/IPropertyMapping.cs
@@ -1,0 +1,13 @@
+﻿// Licensed to Elasticsearch B.V under one or more agreements.
+// Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information.
+
+namespace Elastic.Clients.Elasticsearch
+{
+	public interface IPropertyMapping
+	{
+		string Name { get; }
+
+		bool Ignore { get; }
+	}
+}
